requirements.txt
django
djangorestframework
dotenv
moviepy
drf_spectacular
celery
redis
pytest
pytest-django